refactor : amount 타입 변경, 매직 넘버 상수화
변경한 내용 :
1. amount 타입을 int - > long 으로 변경
-> 구매금액이 21억을 넘을 수도 있을 것 같아서
2. 매직 넘버 전체적으로 상수화
3. test 코드 변경 (타입 변경)

+++ src/main/java/lotto/domain/LottoFactory.java
@@ -16,7 +16,7 @@ public class LottoFactory {
         this.lottoNumberGenerator = lottoNumberGenerator;
     }
 
-    public List<Lotto> createAllLottos(int quantity) {
+    public List<Lotto> createAllLottos(long quantity) {
         List<Lotto> lottos = new ArrayList<>();
         for (int i = 0; i < quantity; i++) {
             Lotto lotto = createSingleLotto();
